i-bot/pyproject.toml
[project]
name = "i-bot"
version = "0.1.0"
dependencies = [
    "langchain",
    "langgraph",
    "langchain-openai",
    "langchain-postgres",
    "psycopg-binary",
    "sqlalchemy",
    "fastapi",
    "uvicorn",
    "python-dotenv",
    "langmem",
    "python-multipart",
    "pypdf",
    "markitdown",
]

[tool.setuptools]
packages = ["src"]
